# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Production ERP Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change current user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # USERS
...
  /entities:
    get:
      operationId: listEntities
      tags: [entities]
      summary: List all entities
      responses:
        "200":
          description: List of entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entity"
    post:
      operationId: createEntity
      tags: [entities]
      summary: Create a new entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityInput"
      responses:
        "201":
          description: Entity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
  /entities/{id}:
    get:
      operationId: getEntity
      tags: [entities]
      summary: Get entity by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entity found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
    put:
      operationId: updateEntity
      tags: [entities]
      summary: Update entity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityUpdate"
      responses:
        "200":
          description: Entity updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
    delete:
      operationId: deleteEntity
      tags: [entities]
      summary: Delete entity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # Multilingual label stored as JSONB {ru: ..., en: ..., he: ...}
    MultilingualText:
      type: object
      properties:
        ru:
          type: string
        en:
          type: string
        he:
          type: string

    # AUTH
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, email, firstName, lastName, roleId, language, direction, isActive]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        roleId:
          type: integer
        roleName:
          $ref: "#/components/schemas/MultilingualText"
        language:
          type: string
          enum: [ru, en, he]
        direction:
          type: string
          enum: [ltr, rtl]
        startPageId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        permissions:
          $ref: "#/components/schemas/RolePermissions"
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
...
    RecordPermission:
      type: object
      required: [view, create, update, delete]
      properties:
        view:
          type: boolean
        create:
          type: boolean
        update:
          type: boolean
        delete:
          type: boolean
    RoleAdminCaps:
      type: object
      required: [pages, entities, roles, users, translations]
      properties:
        pages:
          type: boolean
        entities:
          type: boolean
        roles:
          type: boolean
        users:
          type: boolean
        translations:
          type: boolean
    RolePermissions:
      type: object
      required: [superAdmin, admin, pageIds, records]
      properties:
        superAdmin:
          type: boolean
        admin:
          $ref: "#/components/schemas/RoleAdminCaps"
        pageIds:
          type: array
          items:
            type: integer
        records:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/RecordPermission"
...
    Entity:
      type: object
      required: [id, entityKey, nameJson, icon, sortOrder, isActive, createdAt, updatedAt]
      properties:
        id:
          type: integer
        entityKey:
          type: string
        nameJson:
          $ref: "#/components/schemas/MultilingualText"
        descriptionJson:
          $ref: "#/components/schemas/MultilingualText"
        icon:
          type: string
        pageId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EntityInput:
      type: object
      required: [entityKey, nameJson, icon]
      properties:
        entityKey:
          type: string
        nameJson:
          $ref: "#/components/schemas/MultilingualText"
        descriptionJson:
          $ref: "#/components/schemas/MultilingualText"
        icon:
          type: string
        pageId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
          default: true
    EntityUpdate:
      type: object
      properties:
        entityKey:
          type: string
        nameJson:
          $ref: "#/components/schemas/MultilingualText"
        descriptionJson:
          $ref: "#/components/schemas/MultilingualText"
        icon:
          type: string
        pageId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean

    # FIELDS
...