# lib/api-zod/src/generated/types/recordPermission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Production ERP Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface RecordPermission {
  view: boolean;
  create: boolean;
  update: boolean;
  delete: boolean;
}