# lib/api-zod/src/generated/types/roleAdminCaps.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Production ERP Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface RoleAdminCaps {
  pages: boolean;
  entities: boolean;
  roles: boolean;
  users: boolean;
  translations: boolean;
}