# lib/api-zod/src/generated/types/rolePermissions.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Production ERP Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { RoleAdminCaps } from './roleAdminCaps';
import type { RolePermissionsRecords } from './rolePermissionsRecords';

export interface RolePermissions {
  superAdmin: boolean;
  admin: RoleAdminCaps;
  pageIds: number[];
  records: RolePermissionsRecords;
}