# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Production ERP Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { MultilingualText } from './multilingualText';
import type { RolePermissions } from './rolePermissions';
import type { UserProfileDirection } from './userProfileDirection';
import type { UserProfileLanguage } from './userProfileLanguage';

export interface UserProfile {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  roleId: number;
  roleName?: MultilingualText;
  language: UserProfileLanguage;
  direction: UserProfileDirection;
  /** @nullable */
  startPageId?: number | null;
  isActive: boolean;
  permissions?: RolePermissions;
}